[3859] [New sql update for realmd 3859_spell_affect.sql] 1. Implemented 64bit SpellFamilyFlags 2. Fixed tooltips of spells that are affected by SpellMods

+++ src/game/ObjectMgr.h
@@ -150,7 +150,7 @@ struct SpellAffection
     uint16 Category;
     uint16 SkillId;
     uint8 SpellFamily;
-    uint32 SpellFamilyMask;
+    uint64 SpellFamilyMask;
     uint16 Charges;
 };
